frontend/requirements.txt
streamlit
requests
plotly-express
networkx
pyvis
matplotlib